% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lagmat}
\alias{lagmat}
\title{lagmat}
\usage{
lagmat(x, lags)
}
\arguments{
\item{x}{the series to be lagged}

\item{lags}{number of lags desired}
}
\value{
matrix with dimension \code{[NROW(x),length(lags)]}
}
\description{
Creates a lagged matrix with the desired number of lags.
}
\examples{
x = rnorm(100)
lx <- lagmat(x,2)
tail(lx)
tail(x)
}

