% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mplott}
\alias{mplott}
\title{mplott}
\usage{
mplott(x, wherelegend = "bottomleft", textlegend = colnames(x), main = "",
  return.to.default = T, ...)
}
\arguments{
\item{x}{a matrix to be plotted}

\item{wherelegend}{where to place the legend}

\item{textlegend}{what should the legend read (see details).}

\item{main}{main title of the plot, same as in \code{\link{plot.default}}.}

\item{return.to.default}{for reverting back to previous par settings.
Default is \code{return.to.default=TRUE}}

\item{...}{more graphical parameters can be given as arguments.}
}
\value{
called for its side effect.
}
\description{
Multivariate plot.
}
\details{
Multivariate plot. Limited to 5 series. Legend is added automatically using
\code{colnames(x)}.
}
\seealso{
\code{\link{par}}, \code{\link{plot.default}}.
}

