% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plott}
\alias{plott}
\title{plott}
\usage{
plott(x, y = c(1:length(x)), add = FALSE, pch = 19, xlab = "",
  col = 1, main = NULL, ty = "b", return.to.default = TRUE, ...)
}
\arguments{
\item{x}{series to be plotted.}

\item{y}{possible second series, if provided a scatter plot is created.}

\item{add}{if \code{add = TRUE} the series is added to existing active device. The active device graphical
parameters must match, meaning it must be created using \code{return.to.default=F}.}

\item{pch}{same as in \code{\link{par}}.}

\item{xlab}{a label for the x axis, same as in \code{\link{plot.default}}.}

\item{main}{main title of the plot, same as in \code{\link{plot.default}}.}

\item{col}{Color of the series, same as in \code{\link{par}}.}

\item{ty}{character indicating the type of plotting, any of the types as in \code{\link{plot.default}}.}

\item{return.to.default}{for reverting back to previous par settings.
Default is \code{return.to.default=TRUE}.}

\item{...}{more graphical parameters can be given as arguments.}
}
\value{
Called for its side effect.
}
\description{
Plotting longitudinal series.
}
\details{
Sets default parameters to get a nicer figure. If y is given then a scatter
plot is created. y can also be of class Date. If add=TRUE, add the series to
an existing device. Make a sensible choice as to which series should be
plotted first so that the ylim and xlim cover are sufficient.
}
\seealso{
\code{\link{par}}, \code{\link{plot.default}}
}

