% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tsideplot}
\alias{tsideplot}
\title{tsideplot}
\usage{
tsideplot(series1, series2, main = "", return.to.default = T,
  xaxis = NULL, col = "red", ...)
}
\arguments{
\item{series1,series2}{First and second series to be plotted.}

\item{main}{main title of the plot, same as in \code{\link{plot.default}}.}

\item{return.to.default}{for reverting back to previous par settings.}

\item{xaxis}{Optional, the xaxis to be used, see details.}

\item{col}{Color of the second series, same as in \code{\link{par}}.}

\item{...}{more graphical parameters can be given as arguments.}
}
\value{
Called for its side effect.
}
\description{
Create a plot of two series vertical axes on both left and right side.
}
\details{
Create a plot of two series with y-axes on both left and right side.
\code{Set return.to.default=TRUE} to keep the new settings, otherwise
default to revert to previous par values. xaxis parameter is the optional
xaxis, if not provided then \code{if(is.null(xaxis)) {xaxis=
c(1:length(series1))}} is used.
}
\seealso{
\code{\link{par}}, \code{\link{plot.default}}.
}

